// src/generated/models/Aum_aiinitiativesesModel.ts
export const Aum_aiinitiativesesaum_aisolutiontype = {
  751770000: 'Predictive',
  751770001: 'AgenticAI',
  751770002: 'WorkFlows',
  751770003: 'Analytics'
} as const;
export type Aum_aiinitiativesesaum_aisolutiontype = keyof typeof Aum_aiinitiativesesaum_aisolutiontype;
export const Aum_aiinitiativesesaum_benefittype = {
  751770000: 'Tangible',
  751770001: 'Intangible'
} as const;
export type Aum_aiinitiativesesaum_benefittype = keyof typeof Aum_aiinitiativesesaum_benefittype;
export const Aum_aiinitiativesesaum_businessimpactcategory = {
  751770000: 'Efficiencygain_20_orvalue_AED5M',
  751770001: 'Gain10_20_orvalueAED1_5M',
  751770002: 'Gain_10_orvalue_AED1M'
} as const;
export type Aum_aiinitiativesesaum_businessimpactcategory = keyof typeof Aum_aiinitiativesesaum_businessimpactcategory;
export const Aum_aiinitiativesesaum_demandtype = {
  751770000: 'Turnkey',
  751770001: 'Enhancement'
} as const;
export type Aum_aiinitiativesesaum_demandtype = keyof typeof Aum_aiinitiativesesaum_demandtype;
export const Aum_aiinitiativesesaum_entitytype = {
  751770000: 'Demand',
  751770001: 'UseCase',
  751770002: 'DeploymentProject',
  751770003: 'RetreatIdea',
  751770004: 'AIApplication'
} as const;
export type Aum_aiinitiativesesaum_entitytype = keyof typeof Aum_aiinitiativesesaum_entitytype;
export const Aum_aiinitiativesesaum_feasibilitydatareadiness = {
  751770000: 'Dataavailable_lowsystemcomplexity',
  751770001: 'Partialdata_moderatecomplexity',
  751770002: 'Datagaps_highcomplexity'
} as const;
export type Aum_aiinitiativesesaum_feasibilitydatareadiness = keyof typeof Aum_aiinitiativesesaum_feasibilitydatareadiness;
export const Aum_aiinitiativesesaum_ftecostsavings = {
  751770000: 'ProductivityValue_AED2Mp_a_',
  751770001: 'ProductivityValueAED0_5M_2Mp_a_',
  751770002: 'ProductivityValue_AED0_5Mp_a_'
} as const;
export type Aum_aiinitiativesesaum_ftecostsavings = keyof typeof Aum_aiinitiativesesaum_ftecostsavings;
export const Aum_aiinitiativesesaum_priority = {
  751770000: 'High',
  751770001: 'Medium',
  751770002: 'Low'
} as const;
export type Aum_aiinitiativesesaum_priority = keyof typeof Aum_aiinitiativesesaum_priority;
export const Aum_aiinitiativesesaum_programproject = {
  751770000: 'Program',
  751770001: 'Project'
} as const;
export type Aum_aiinitiativesesaum_programproject = keyof typeof Aum_aiinitiativesesaum_programproject;
export const Aum_aiinitiativesesaum_source = {
  751770000: 'D2D',
  751770001: 'AIDeployment',
  751770002: 'AlHasbah',
  751770003: 'AIRetreat',
  751770004: 'StateofAI'
} as const;
export type Aum_aiinitiativesesaum_source = keyof typeof Aum_aiinitiativesesaum_source;
export const Aum_aiinitiativesesaum_status = {
  751770000: 'Planning',
  751770001: 'InDiscovery',
  751770002: 'Active',
  751770003: 'Pilot',
  751770004: 'Deployed',
  751770005: 'OnHold',
  751770006: 'Completed',
  751770007: 'Cancelled'
} as const;
export type Aum_aiinitiativesesaum_status = keyof typeof Aum_aiinitiativesesaum_status;
export const Aum_aiinitiativesesaum_strategicalignment = {
  751770000: 'Directlysupports2_DEWAstrategicpillarsorUAEAI2031',
  751770001: 'Supports1strategicpillar',
  751770002: 'Minororindirectalignment'
} as const;
export type Aum_aiinitiativesesaum_strategicalignment = keyof typeof Aum_aiinitiativesesaum_strategicalignment;
export const Aum_aiinitiativesesaum_timetofirstvalue = {
  751770000: 'Firstvaluewithin12months',
  751770001: 'Firstvaluein12_24months',
  751770002: 'Firstvalueafter24months'
} as const;
export type Aum_aiinitiativesesaum_timetofirstvalue = keyof typeof Aum_aiinitiativesesaum_timetofirstvalue;
export const Aum_aiinitiativesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Aum_aiinitiativesesstatecode = keyof typeof Aum_aiinitiativesesstatecode;
export const Aum_aiinitiativesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Aum_aiinitiativesesstatuscode = keyof typeof Aum_aiinitiativesesstatuscode;

export interface Aum_aiinitiativesesBase {
  aum_agenticengineer?: string;
  aum_agentid?: string;
  aum_aiinitiativesid: string;
  aum_aisolutiontype?: Aum_aiinitiativesesaum_aisolutiontype;
  aum_aitechnology?: string;
  aum_annualvolume?: number;
  aum_benefittype?: Aum_aiinitiativesesaum_benefittype;
  aum_businessimpactcategory?: Aum_aiinitiativesesaum_businessimpactcategory;
  aum_businessimpactdetails?: string;
  aum_costincurred?: number;
  aum_costsaving?: number;
  aum_datedeployed?: string;
  aum_dateofsubmission?: string;
  aum_demandno?: string;
  aum_demandtype?: Aum_aiinitiativesesaum_demandtype;
  aum_department?: string;
  aum_division?: string;
  aum_divisionheadreview?: boolean;
  aum_domain?: string;
  aum_employeesimpactedcount?: string;
  aum_entitytype?: Aum_aiinitiativesesaum_entitytype;
  aum_estimatedtimesavingshoursmonth?: string;
  aum_feasibilitydatareadiness?: Aum_aiinitiativesesaum_feasibilitydatareadiness;
  aum_ftecostsavings?: Aum_aiinitiativesesaum_ftecostsavings;
  aum_implementationmandays?: string;
  aum_initiativedescription?: string;
  aum_initiativeowner?: string;
  aum_initiativetitle: string;
  aum_keymilestones?: string;
  aum_kpissuccessmetrics?: string;
  aum_priority?: Aum_aiinitiativesesaum_priority;
  aum_problemstatement?: string;
  aum_productivitygain?: string;
  aum_programproject?: Aum_aiinitiativesesaum_programproject;
  aum_progresspct?: number;
  aum_proposedaisolution?: string;
  aum_reportingyear?: number;
  aum_risksdependencies?: string;
  aum_source?: Aum_aiinitiativesesaum_source;
  aum_sourcemaster?: string;
  aum_stage?: string;
  aum_startdate?: string;
  aum_status?: Aum_aiinitiativesesaum_status;
  aum_strategicalignment?: Aum_aiinitiativesesaum_strategicalignment;
  aum_supportingevidence?: string;
  aum_systemsintegrated?: string;
  aum_targetcompletion?: string;
  aum_technology?: string;
  aum_timetofirstvalue?: Aum_aiinitiativesesaum_timetofirstvalue;
  aum_turnkeyproject?: boolean;
  aum_uid?: string;
  aum_useradoption?: string;
  cr978_aisolutiontypemaster?: string;
  cr978_inactivereason?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Aum_aiinitiativesesstatecode;
  statuscode?: Aum_aiinitiativesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Aum_aiinitiativeses extends Aum_aiinitiativesesBase {
  aum_aisolutiontypename?: string;
  aum_benefittypename?: string;
  aum_businessimpactcategoryname?: string;
  aum_demandtypename?: string;
  aum_divisionheadreviewname?: string;
  aum_entitytypename?: string;
  aum_feasibilitydatareadinessname?: string;
  aum_ftecostsavingsname?: string;
  aum_priorityname?: string;
  aum_programprojectname?: string;
  aum_sourcename?: string;
  aum_statusname?: string;
  aum_strategicalignmentname?: string;
  aum_timetofirstvaluename?: string;
  aum_turnkeyprojectname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}